Redirect active users to journal
If you are a revisiting user and are actively using the Journal, you probably don't to see the index every time.
The bit of code change says that if you are landing at the root of the path and there is a profile, redirect to the journal. If you visit the index.html or any other .html, do not redirect to the Journal.